move find_project_root to right place and add looks_like_project_root
